+++ outputs.tf
@@ -1,3 +1,13 @@
+output "account_id" {
+  description = "Account ID"
+  value       = aws_organizations_account.account.id
+}
+
+output "account_name" {
+  description = "Account name"
+  value       = var.account_name
+}
+
 output "arn" {
   value = module.iam_role_iam_manager.arn
 }